apps/api/pyproject.toml
[project]
name = "rag-pipeline-api"
version = "0.1.0"
description = "RAG Pipeline FastAPI Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.8.0",
    "celery[redis]>=5.6.3",
    "redis>=6.2.0",
    "httpx>=0.28.0",
    "websockets>=15.0",
    "python-multipart>=0.0.20",
    "structlog>=25.0.0",
    "markitdown>=0.1.5",
    "beautifulsoup4>=4.14.0",
    "playwright>=1.58.0",
    "trafilatura>=2.0.0",
    "langchain-anthropic>=0.4.0",
    "langchain-core>=0.3.0",
    "langgraph>=1.1.0",
    "langchain>=1.2.0",
    "langchain-openai>=0.3.0",
    "pydantic-ai>=0.1.0",
    "numpy>=2.0.0",
    "a2a-sdk",
    "fastembed>=0.8.0,<1.0.0",
    "tiktoken>=0.12.0,<1.0.0",
    "qdrant-client>=1.17.1,<2.0.0",
    "mcp>=1.27.0,<2.0.0",
    "opentelemetry-api>=1.33.0,<2.0.0",
    "opentelemetry-sdk>=1.33.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.54b0",
    "opentelemetry-instrumentation-celery>=0.54b0",
    "opentelemetry-instrumentation-httpx>=0.54b0",
    "opentelemetry-exporter-otlp>=1.33.0,<2.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0,<8.0.0",
    "python-jose[cryptography]>=3.4.0,<4.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "langsmith>=0.3.42,<1.0.0",
    "sentry-sdk[fastapi]>=2.29.1,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]